+++ tests/testthat/test_importPolarized.r
@@ -18,6 +18,27 @@ test_that("error on arguments in importPolarized", {
     ),
     regexp = "contains fewer individuals"
   )
+  
+    expect_error(
+    object = importPolarized(
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
+      changePolarity = c(0, 1, 0),
+      ChosenInds = 1:6,
+      ChosenSites = "all"
+    ),
+    regexp = "must be a logical vector or a list of logical vectors"
+  )
+
+  expect_error(
+    object = importPolarized(
+      files = system.file("extdata", "data7x3.txt", package = "diemr"),
+      changePolarity = list(c(TRUE, FALSE, "x")),
+      ChosenInds = 1:6,
+      ChosenSites = "all"
+    ),
+    regexp = "must be a logical vector or a list of logical vectors"
+  )
+
 })